Simplified code and modified profile details
