+++ meeting_client/app/src/main/res/values/strings.xml
@@ -15,8 +15,7 @@
     <string name="title_activity_add_users_to_meeting">AddUserstoMeeting</string>
 
     <!--registration/login strings-->
-    <string name="first_name">First Name</string>
-    <string name="last_name">Last Name</string>
+    <string name="name">Name</string>
     <string name="user_name">Username</string>
     <string name="email">Email</string>
     <string name="password">Password</string>
